Fix: changed logic on delete if only 1 page is left in resource category

+++ routes/resourcePages.js
@@ -1,6 +1,5 @@
 const express = require('express');
 const router = express.Router();
-const _ = require('lodash')
 
 // Import middleware
 const { attachRouteHandlerWrapper } = require('../middleware/routeHandler')
@@ -105,13 +104,13 @@ async function deleteResourcePage(req, res, next) {
   const IsomerFile = new File(accessToken, siteName)
   const resourcePageType = new ResourcePageType(resourceRoomName, resourceName)
   IsomerFile.setFileType(resourcePageType)
-  await IsomerFile.delete(pageName, sha)
-
-  // Check if collection has any pages left, and delete if none left
   const resources = await IsomerFile.list()
-  if (_.isEmpty(resources)) {
+  if (resources.length === 1) {
+    // If there is only 1 page left, we can delete the entire category
     const IsomerResource = new Resource(accessToken, siteName)
     await IsomerResource.delete(resourceRoomName, resourceName)
+  } else {
+    await IsomerFile.delete(pageName, sha)
   }
 
   res.status(200).send('OK')
